+++ src/main/java/subway/domain/core/AddStationToRightStrategy.java
@@ -1,12 +1,9 @@
-package subway.domain.strategy;
+package subway.domain.core;
 
-import static subway.domain.Direction.LEFT;
+import static subway.domain.core.Direction.LEFT;
 
 import java.util.List;
 import java.util.Optional;
-import subway.domain.Distance;
-import subway.domain.Section;
-import subway.domain.Station;
 
 public class AddStationToRightStrategy implements AddStationStrategy {
     @Override
