+++ src/main/java/subway/domain/core/Station.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.core;
 
 import java.util.Objects;
