+++ src/main/java/subway/service/LineService.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.domain.Line;
+import subway.domain.core.Line;
 import subway.dto.LineAddRequest;
 import subway.dto.LineResponse;
 import subway.dto.LineUpdateRequest;
