+++ src/test/java/subway/controller/LineControllerTest.java
@@ -12,8 +12,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.common.IntegrationTest;
-import subway.domain.Line;
-import subway.domain.Section;
+import subway.domain.core.Line;
+import subway.domain.core.Section;
 import subway.dto.LineAddRequest;
 import subway.dto.LineResponse;
 import subway.dto.LineUpdateRequest;
