+++ src/test/java/subway/domain/core/AddStationToRightStrategyTest.java
@@ -1,4 +1,4 @@
-package subway.domain.strategy;
+package subway.domain.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,9 +7,6 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import subway.domain.Distance;
-import subway.domain.Section;
-import subway.domain.Station;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SuppressWarnings("NonAsciiCharacters")
