+++ src/test/java/subway/domain/core/StationTest.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.core;
 
 import static org.assertj.core.api.Assertions.assertThat;
