+++ src/test/java/subway/repository/LineMapperTest.java
@@ -6,10 +6,10 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-import subway.domain.Distance;
-import subway.domain.Line;
-import subway.domain.Section;
-import subway.domain.Station;
+import subway.domain.core.Distance;
+import subway.domain.core.Line;
+import subway.domain.core.Section;
+import subway.domain.core.Station;
 import subway.entity.LineEntity;
 import subway.entity.SectionEntity;
 import subway.entity.StationEntity;
