+++ src/test/java/subway/service/LineServiceTest.java
@@ -12,8 +12,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-import subway.domain.Line;
-import subway.domain.Section;
+import subway.domain.core.Line;
+import subway.domain.core.Section;
 import subway.dto.LineAddRequest;
 import subway.dto.LineResponse;
 import subway.dto.LineUpdateRequest;
